% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNorm.R
\name{checkNormality}
\alias{checkNormality}
\title{Perform all Normality Tests and Visualizations}
\usage{
checkNormality(data, var, line_color, ...)
}
\arguments{
\item{data}{A data frame containing a variable of interest.}

\item{var}{The name of the numeric variable in the dataset to be checked for normality (typically a dependent variable).}

\item{line_color}{A character string specifying the color of the normal distribution curve on the histogram and QQ plot that the function will output.}

\item{...}{Additional arguments passed to underlying functions}
}
\value{
A message containing the results of the Shapiro-Wilk test for normality and the combined histogram and QQ plot of the variable of interest.
}
\description{
This function takes in a numeric variable from a data frame and uses it to run a normality test.
At the same time, it outputs a histogram with a normal distribution curve, a QQ plot, and a Shapiro-Wilk test for normality using the input variable.
}
\examples{
df <- data.frame(outcome = rnorm(100))
checkNormality(df, outcome, line_color = "red")

}
