% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Compute All Important Statistics}
\usage{
stats(x)
}
\arguments{
\item{x}{a numeric vector for which to compute key statistics.}
}
\value{
A formatted table displaying the computed statistics:
\item{Mean}{The average of the values.}
\item{Variance}{The variance of the values.}
\item{StDev}{The standard deviation of the values.}
\item{Min}{The minimum value in the vector.}
\item{Max}{The maximum value in the vector.}
\item{Median}{The median of the values.}
\item{25th Percentile}{The first quartile (Q1) of the values.}
\item{75th Percentile}{The third quartile (Q3) of the values.}
\item{Coefficient_of_Variation}{The standard deviation divided by the mean, representing relative variability.}
}
\description{
This function computes key descriptive statistics for a numeric vector, including
measures of central tendency, variability, and percentiles.
These results can inform a user's choice of statistical analysis, especially for those that require information on dispersion.
}
\examples{
data <- c(1, 10, 5, 3, 7, 18)
stats(data)

}
